Documented the secure storage service
